[FIX] project_enterprise: ensure test does not depend on demo data
Before this commit, the `ProjectEnterpriseTestUi.test_01_ui` test fails
without demo data because the user is not considered as overloaded. The
reason is because the user has more than one task assigned to him when
the demo data is installed and so when a new task is planned for him,
the system will see the sum of the allocated_hours of all tasks assigned
to him inside the period displayed in the gantt is higher than the hours
he can do.

This commit makes sure the test does not depend on the demo data and
increase the allocated hours of the task creates inside the test.

runbot-error-111415

closes odoo/enterprise#82987

X-original-commit: 36c3552baff426dde26ebee352a3d08655a2ee1b
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>