[FIX] account_bank_statement_extract: activate currency in order to validate document
If EUR is not active, `check_all_status` fails (because
`AccountMove._post` fails).

Before #80180, EUR was implicitly enabled by the multi-company setup
of MailCommon: it creates a company with the currency EUR, and
`ResCompany.create` automatically activates the company
currency. Since the dependency on `MailCommon` was removed from
`TestBankStatementExtractProcess`.

This is not visible on the runbot, because it always installs l10n_be
which automatically enables EUR.

closes odoo/enterprise#80361

Signed-off-by: Xavier Morel (xmo) <xmo@odoo.com>