[FIX] partner_commission: add PO lines to existing RFQ if referrer is set on SO
Before this commit:
When a (SO) with subcontracted service product is confirmed, a (PO) is created.
However, if a second SO is created with referrer for the same vendor, a new PO
is generated, even if an RFQ already exists for that vendor.

After this commit:
When confirming an SO with a referrer and a subcontracted service product , and
if an RFQ already exists for the same vendor, the PO lines are added to the
existing RFQ instead of creating a new PO.

task-4578812

closes odoo/enterprise#81506

X-original-commit: edd3a01979bc3f71d44a26d2ea4206b48317bf82
Related: odoo/odoo#201877
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>