[FIX] sale_subscription: prevent payment link with renewed orders
Before this commit, it was possible to pay payment links linked to
renewed orders. It would cause issues as the renewed order would be
reopened once the transaction was set to done. Two subscription in
progress would live side by side. This commit ensure that such links
can't be created and existing links can't be used.

closes odoo/enterprise#82308

Taskid: 4607315
X-original-commit: 4e4ce5e20020b00767a0c58821a8699ecf2e74f5
Signed-off-by: Thibault Francois <tfr@odoo.com>
Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>