[FIX] account_accountant: manual operations
This commit is a partial backport of:
https://github.com/odoo/enterprise/commit/eaf8361f44208346c20e4c50c2d9914ab2d4f3f5
In the commit, we decided to revamp the manual operation tab in terms of UI by
moving some divs and did minor clean up to make the code more readable.

Moreover, we removed a lot of extra margin that would make the view less
readable.

task:4542756

closes #79233

closes odoo/enterprise#80552

X-original-commit: 4902236572309080f3561073d5bb63f275ecfc47
Signed-off-by: Maximilien La Barre (malb) <malb@odoo.com>
Signed-off-by: Florian Gilbert (flg) <flg@odoo.com>