[FIX] web_grid: show all records on mobile
Steps to reproduce
==================

- Install hr_timesheets
- Go to Timesheets > All Timesheets
- Use the week scale
- Scroll to the bottom
=> Take note of the last record (e.g.: Walter Horton)
- Now use a mobile viewport
- Make sure the week scale is still applied
- Scroll to the bottom
=> The last record is not the same one as the desktop view

Cause of the issue
==================

The grid view uses a virtual scroll hook in order to not have to many
dom nodes in memory at the same time. For this to work, we need a
consistent row height.

On mobile, when grouping by x amount of fields, they are stacked on top
of each other. When grouping by more than 2 fields, there wasn't enough
space to display them all.
The `grid-template-rows` was set to auto on mobile in order to
accomodate that case.

Using auto means we no longer have a consistent row height.

Solution
========

We restore the `gridTemplateRows` and allocate extra space when needed.

opw-4208357

closes odoo/enterprise#81797

X-original-commit: d5d1b7a07f69a3267be172b2638d331872b60dbc
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>