[IMP] stock_barcode: make packaging unit the main line unit
This commit changes the default behavior of immediate transfers created
from Barcode app if the user scanned a packaging barcode. Previously,
the line unit would always be the product base unit and the quantity is
determined based on the factor of the scanned packaging. After this
commit, the line unit would be the scanned packaging directly with a
quantity of one.

Task-4471379

closes odoo/enterprise#84676

Related: odoo/odoo#208431
Signed-off-by: Arnold Moyaux (arm) <arm@odoo.com>
Signed-off-by: Mohammad Abdulmoneim (abdu) <abdu@odoo.com>