[FIX] stock_barcode: add test for GS1 barcode addition & removal
Currently, an exception is generated when a user tries to save a product record
without a barcode, whether it is a newly created record or an existing one.

Steps to reproduce:

1. Install the Inventory module.
2. Go to Inventory settings.
3. Set Barcode Nomenclature to Default GS1 Nomenclature.
4. Go to Inventory -> Products, create a product with a barcode, and save it.
5. Edit the same product, remove the barcode, and save it again.
6. An error occurs.

Error: `StopIteration`

This issue [1] occurs because when a user tries to save the record without a
barcode, the value is false, and an false value is not iterable.

[1] - https://github.com/odoo/odoo/blob/9eb6c16cd7d4fe9364d043e2af96876bf82a44d4/addons/barcodes_gs1_nomenclature/models/barcode_nomenclature.py#L163

This fix resolves an issue where GS1 preprocessing was always applied when checking
for duplicate product barcodes. By introducing the skip_preprocess_gs1 context flag,
the GS1 preprocessing step is now bypassed when verifying barcode uniqueness.

sentry-6267124218

closes odoo/enterprise#83325

X-original-commit: be8eedde998dce01ab1125870ea77cdd534ea314
Related: odoo/odoo#205748
Signed-off-by: Steve Van Essche <svs@odoo.com>
Signed-off-by: Kishan Kumar Jaiswal (kkuj) <kkuj@odoo.com>