[IMP] hr_payroll_account_iso20022: remove IBAN check for consistency
When trusting a bank account from the employee's form view, a check
was performed to verify that the bank account has a valid IBAN.
This check is not performed on the bank account form view as a bank
account (e.g. in the US) does not necessarily follow the IBAN format.
This commits remove the IBAN validity check on the employee form view
for consistency with the bank account view.

task-4613593

closes odoo/enterprise#80569

Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>