[FIX] knowledge: set email on admin during UI tests
The removal of `MailCommon` in
https://github.com/odoo/enterprise/pull/80180 for
`TestKnowledgeUICommon` had the unintended side effect of removing the
email from the user `admin`.
It happens that tours using that class actually send emails and thus
require the email address to be set on the user, which is not the case
in no demo runs.

Runbot Error 145750

closes odoo/enterprise#80570

Signed-off-by: William Braeckman (wbr) <wbr@odoo.com>