[IMP] account_accountant: Hide draft lines when no partners on the statement
When no partner is set on the bank transaction, we do not show the draft
items among candidates.

Reason : avoid showing too many irrelevant candidates quickly making the
feature unusable.

But if the user filters the candidate and selects a partner, then show
the draft items of this partner among the candidates. Even though the
partner filter is a conscious one (input by the user) and not computed
by the partner on the bank transaction.

This is done by adding a 'posted' filter to the list view that is
automatically added when no partner is set in the statement.

task-4554569

closes odoo/enterprise#81913

X-original-commit: bebfc6ae42aa7291cf132dfdbbdb566fe9fd27f2
Signed-off-by: William André (wan) <wan@odoo.com>
Signed-off-by: Antoine Boonen (aboo) <aboo@odoo.com>