[IMP] documents: use res_model=False when not linked
Purpose
=======
When grouping documents by models, we sometimes get `False`, sometimes `Documents` when
the document is not linked to a record. We would like to fix this inconsistency.

Use False instead of 'documents.document' when a document is not linked to a record.

Attachment still have 'documents.document' if not linked to a record, so users see
the attachment in the technical view if they have access to the document (otherwise
they won't see it).

Remove dead code as well.

Task-4532059

closes odoo/enterprise#80653

Related: odoo/upgrade#7319
Signed-off-by: Florian Charlier (flch) <flch@odoo.com>