[FIX] l10n_be_hr_*: fix choose bike behavior
When an existing bike is selected when signing a new contract, the
future driver is not set on the biked, so it is not reserved and can be
taken by multiple people.

Also when you unselect 'transport_mode_bike' on an existing contract,
the bike_id is not removed (but invisible), which leads to prevent any
person with this kind of contract to select a bike in an offer. (same
behavior with new bike, the model is not removed)

closes odoo/enterprise#80746

X-original-commit: c8b1f3c6dd5a97bd02776bb33024ac38fb9cbe5f
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>
Signed-off-by: Pierre Masereel (pim) <pim@odoo.com>