[FIX] l10n_mx_edi_pos: fix refund crash when the order is invoiced
Steps to reproduce
==================
- Install PoS, and select a Mexican company
- Create an order O1, and in the payment screen, select "Invoice" and
  confirm the dialog
- Validate the order
- Click on "Orders" and set the filter to "Paid" to see order O1
- Click O1, click "Refund", and from the product screen, click "Payment"

Observed behavior: White screen with traceback in the console.

Issue
=====
It's happening because in l10n_mx_edi_pos, an invoiced order is supposed
to have fields `l10n_mx_edi_cfdi_to_public` and `l10n_mx_edi_usage` [1],
and they are being accessed in the `product_screen.xml` [2]. That works
for the original ordre (aka O1), however, when the refund order is
created, we set its `to_invoice` to `true` if the original order is
invoiced [3], but we don't set any values to the fields
`l10n_mx_edi_cfdi_to_public` and `l10n_mx_edi_usage`, causing the crash
in the payment screen [2].

Fix
===
If the refund order is invoiced, we set the `l10n_mx_edi_usage` to "G02"
which corresponds to 'Returns, discounts or bonuses' (assuming that a
refund can be seen as a 'return' here). We also copy the field
`l10n_mx_edi_cfdi_to_public` from the original order.

[1]: https://github.com/odoo/enterprise/blob/a0a97f2c87f76ae035394a67b09d71ec8f4567e7/l10n_mx_edi_pos/static/src/app/screens/payment_screen/payment_screen.js#L18-L21
[2]: https://github.com/odoo/enterprise/blob/0396d4f0ac33640dc042bb97e9e8d4b498e6a128/l10n_mx_edi_pos/static/src/app/screens/payment_screen/payment_screen.xml#L7-L12
[3]: https://github.com/odoo/odoo/blob/53e7c0adf093dc950a42989a56b4594f734c17b2/addons/point_of_sale/static/src/app/screens/payment_screen/payment_screen.js#L69-L75

opw-4608630

closes odoo/enterprise#81667

X-original-commit: 97796842c8aaff4cf569581c407b70e39c497a97
Signed-off-by: Joseph Caburnay (jcb) <jcb@odoo.com>
Signed-off-by: Hadi El Yakhni (hael) <hael@odoo.com>