[IMP] appointment: allow remote embedding of form
Remove CSRF check on appointment forms if there is no session or no associated
user for the session.

We consider this safe as we already accept logged-out creation of appointment
and do not modify any existing data so long as no user is logged in. In that way
it is similar to regular form snippets which already work this way since [1]

This greatly eases embedding this form on existing websites for users who do not
wish to use our website integration immediately but still want the benefits of
the app and the ease of use of the form page.

[1]: odoo/odoo@b192d8d23f83b35c21b13f6e4516c18897b80201

task-4593880

closes odoo/enterprise#81684

Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>