[FIX] website_helpdesk_knowledge: make sure tour does not depend on demo data
Before this commit, the `helpdesk_pick_template_as_message_from_knowledge` tour
fails in no demo because the admin user is not named `Mitchell Admin`
and so the step does not find the signature field.

This commit adds the signature on the admin user to make sure the
signature will be the same with demo data and without demo data.

closes odoo/enterprise#81001

X-original-commit: 0477753ca7ae99567db87a9387f3b1781fb13411
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>
Signed-off-by: William Braeckman (wbr) <wbr@odoo.com>