[FIX] whatsapp: clone template attachment to prevent deletion
Problem:

The WhatsApp template's attachment is set on the mail message of the related
WhatsApp message. If the user deletes the mail message's attachment, the
template's attachment also gets deleted.

Solution:

Instead of directly linking the template attachment, the attachment will be
cloned with the appropriate `res_id` and `res_model` before sending. This
ensures that the original template attachment remains intact even if the message
attachment is removed.

Task-4626855

closes odoo/enterprise#81805

X-original-commit: 2273b67ef28b35f90eff12635557840744e87804
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>