[FIX] account_iso20022: allow non IBAN for ISO 20022
We support sending credit transfers using ISO 20022 to accommodate non
SEPA bank transfers (any party having no IBAN number).

However when generating the XML file, we assumed that our company always
had an IBAN number and put it in an IBAN node. For non IBAN numbers
however we need to put it in another node. This commit fixes that.

opw-4585994

closes odoo/enterprise#81418

X-original-commit: 98dbbd355b1d8fbd196cb5cfe2ed73a1434e5326
Signed-off-by: Hugo Poncelet (hupo) <hupo@odoo.com>
Signed-off-by: Dylan Kiss (dyki) <dyki@odoo.com>