[FIX] account_accountant: demo data in python instead of xml
Upgrading from 16 causes issues with demo data.
It happens when the main company has no chart template, because it won't find the correct account.journal for the moves (and account.account for the lines).

To reproduce:

1. Initialize an Odoo 16.0 database with account_accountant and NO demo data. (There will be no country_id set on the default company, so the post_install hook in account won't install l10n_generic_coa)
2. Install demo data (the post_install hook in account is then not fired, still no l10n_generic_coa)
3. Upgrade

By creating the demo data in Python instead of the XML, we can put the condition to avoid creating the move if we do not have a chart template on the company.

opw-4781045

closes odoo/enterprise#87241

X-original-commit: 0b92d0d5cec4d3a21ad98d06c7f4c0c9deb65dac
Signed-off-by: Paolo Gatti (pgi) <pgi@odoo.com>
Signed-off-by: Ricardo Gomes Rodrigues (rigr) <rigr@odoo.com>