[FIX] hr_payroll: fix wrong record for tz computation
The bug has been introduced by this commit : https://github.com/odoo/enterprise/commit/4cba2fb58560a95d5ce93eb6e6c1483ef0428ba5

The bug appears when trying to generate batch of payslips with fully flexible working schedules.
When the contract linked to the payslip has no working schedule, the check of the timezone is
done on the employee linked to the payslip.
The fix only consists on taking the employee of each slip instead of taking the employee on self,
which was causing a 'singleton expected' error.

This commit also contains a test for the related fix in the following
related PR: https://github.com/odoo/odoo/pull/200379

task-4623219

closes odoo/enterprise#81980

X-original-commit: 08fd74f9919eefa36409b0eb5363d46323521f5f
Related: odoo/odoo#203068
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>