[IMP] web_gantt: undo last drag & drop action
When using drag & drop on a pill in the gantt view, we make a notification appear containing an "Undo" button.
Clicking "Undo" will revert the last action made by the drag & drop (can be a rescheduling or a copy of the record).
We also allow the user to be able to "Undo" the change after resizing a pill.

task-4613127

closes odoo/enterprise#80801

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>