[FIX] knowledge: fix error while pasting in the composer
Steps to Reproduce
===============
1. Go to the knowledge article
2. Insert a comment for any text
3. Paste an attachment

It raises a server error

Technical
===============
As for the knowledge first comment's attachments, we patch the uploadData of
attachment uploader hook. But the paste flow directly calls the uploadFiles.

After this commit, we patch the uploadFiles to create thread before the
attachments are uploaded.

Follow-up fix of https://github.com/odoo/enterprise/pull/74871

Task-4331677

closes odoo/enterprise#80841

Signed-off-by: Damien Abeloos (abd) <abd@odoo.com>