[FIX] hr_expense_extract: click on try a sample receipt
This error occurs when we click on 'try a sample receipt' in the Expense module.

Steps to reproduce:
- Install 'hr_expense_extract' module(without demo)
- Expenses > 'try a sample receipt'

Traceback:
AttributeError: 'hr.expense' object has no attribute '_default_journal_id'

At [1], we are encountering this error because the '_default_journal_id' method
is defined in 'hr.expense.post.wizard', but we are attempting to access it from
'hr.expense'.

https://github.com/odoo/enterprise/blob/e10f9aa43ebc4368050c4cb904e81253fc324bdc/hr_expense_extract/wizard/expense_sample_receipt.py#L51

sentry-6362357345

closes odoo/enterprise#82327

X-original-commit: fc10d0da0e5593c3a7a680ed7793baf804722987
Signed-off-by: Julien Alardot (jual) <jual@odoo.com>
Signed-off-by: Aayush Modi (aamo) <aamo@odoo.com>