[IMP] appointment: multiday slots support
Before we only supported appointment slot for one full day.
With the possibility to create slot encompassing multiple days,
we improve the info displayed for the slot to the user.
If a "allday" slot encompassed only one day we still display
"All day" for the slot.

The other allday slots (multi-days) display the days as:
"Mon 4 - Tue 5".
The timezone chosen don't impact allday slots. They are only
based on the timezone of the appointment type.

In the other case of multi-days slot, they are displayed like
alldays slots like before.

Additionally the booking an allday slot create an allday meeting
in backend.

UPG PR: odoo/upgrade#8323

task-2822534

closes odoo/enterprise#81734

Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>