[FIX] l10n_co_dian: prevent Negative Values in DIAN XML
Reproducing the issue in the database currently requires a real digital DIAN certificate. However, it can be tested in Python:

- Create a credit note for a CO company with a foreign currency.
- Generate the XML report.

The tags SourceCurrencyBaseRate and CalculationRate are being added with a negative sign, but the DIAN only accepts positive values.

This commit replace the localy computed rate with the account_move invoice_currency_rate (which meant correcting a few already existing test because of decimal accuracy differences)

opw-4608444

closes odoo/enterprise#82232

X-original-commit: c5715124a5a702f6872dcef04751b42d80015433
Signed-off-by: Antoine Dupuis (andu) <andu@odoo.com>
Signed-off-by: Guillaume Teboul-Tornezy (gute) <gute@odoo.com>