[FIX] mrp_mps: inventory/mrp user can change mps display rows
**Current behavior:**
A user who can otherwise use the entire mps can not add 'rows'
to the view (i.e., write on `ResCompany` `mrp_mps` fields).

**Expected behavior:**
Can add rows to the MPS.

**Steps to reproduce:**
1. Install `mrp_mps` with demo data

2. Edit "Demo" user access rights so they only have
the mrp manager group

3. Open MPS -> click `Rows` dropdown (under search bar) and
add/remove one -> access error

**Cause of the issue:**
Field is on `ResCompany` and the MPS user can't write to that
model.

**Fix:**
Permit writes on `ResCompany` `mrp_mps` fields if they are on an
mps field and the user has the mrp manager group.

opw-4591655

closes odoo/enterprise#87897

X-original-commit: 081d73a79798b11ac4436b0446698bce77860aac
Signed-off-by: Steve Van Essche <svs@odoo.com>
Signed-off-by: Vincent Ethan <etvi@odoo.com>