[IMP] account_sepa_direct_debit: display partners no valid sdd mandate
When users try to register one or multiple payments with 'SEPA Direct
Debit' method, and one or more partners don't have a valid SEPA mandate,
a banner alerts the user and mention the names of the concerned
partners. This makes it easier for the user because he directly sees the
partners for which he has to set a SEPA mandate.

task-4575537

closes odoo/enterprise#89084

X-original-commit: 6513dd694385083da601ebc511cfd4a82906da8f
Related: odoo/upgrade#7977
Signed-off-by: Habib Ayob (ayh) <ayh@odoo.com>