[FIX] account_accountant: deferred move unlink with audit trail
Steps to reproduce:
1. Activate audit trail.
2. Make a customer invoice with start and end date [end date should be in future months], and post.
3. Reset the invoice to draft
Result:
All deferred entries are reversed with new entries at the date of the reset to draft.
Unposted deferred entries are still there with auto post At date.

Cause & fix:
Audit trail alone doesn't prevent the moves from being deleted if they were not posted before, so the check shouldn't be only on the company_id.check_account_audit_trail, but also we should consider move.posted_before and if not, it should be deleted not reversed.
This check is offered by this function _is_protected_by_audit_trail().

closes odoo/enterprise#81241

X-original-commit: 76cbf19aca47bce77eba24679cb38383601982b2
Signed-off-by: John Laterre (jol) <jol@odoo.com>