[FIX] website_documents: don't throw AccessError in sudo mode
Versions
--------
- 18.0+

Steps
-----
1. Have a second company & select it;
2. create a project w/out company_id;
3. set its documents folder to "Projects";
4. create a service product that creates a project on confirm;
5. set created project as its project template;
6. add service product to a quotation;
7. confirm quotation.

Issue
-----
Access Error.

Cause
-----
Commit ac051d00e96c added a constraint on the website_id of documents:
- OR the document's company_id has to match the website's company_id;
- OR the website's company_id has to be in `self.env.companies`.

The Projects folder was created without `website_id`, but was assigned
company 1's website by the `_compute_website_id` method.

When a document gets created by the sale order, it inherits the
`website_id` of the Projects folder.

Hence having a shared folder between companies can lead to access
errors if the website's `company_id` isn't selected in the company
selector.

Solution
--------
Bypass the constraint when in sudo mode, so you don't have to enable the
website's company in the company selector when setting up a sale order.

saas-18.1+:
As constraints are now always checked in sudo mode, move the check to
the `create` & `write` methods.

opw-4548619

closes odoo/enterprise#82002

X-original-commit: 14abb76e09d1cff6523466abd42224fa66a3b276
Signed-off-by: Florian Charlier (flch) <flch@odoo.com>
Signed-off-by: Levi Siuzdak <sile@odoo.com>