[IMP] l10n_uy_edi: support English and dedicated addenda in report
This fixes two issues with the report. The first one is automatically
requesting an English report if the customer's preferred language is
not Spanish. This way the lang field on the partner is somewhat
respected for these externally generated reports too.

The second fix requests a dedicated addenda page when appropriate. By
default, the addenda (e.g. terms and conditions) is added in a small
box at the bottom of the standard PDF report. This can only accomodate
roughly 6 lines of 140 characters. If the addenda exceeds that, the
remainder is silently cut off which is problematic for mandatory
disclosures etc. It's possible to request a dedicated addenda page in
the PDF that fixes this issue. We considered always requesting it, but
it looks strange to have a whole page in the PDF if the addenda is
very short. We therefore make a best effort attempt to figure out if
the addenda will fit in the standard report or not.

task-4750717

closes odoo/enterprise#85796

X-original-commit: b15eba4905ea720b611121efb3e57f299b99e85e
Signed-off-by: Laurent Smet (las) <las@odoo.com>