[IMP] l10n_uy_edi: send line.name as NomItem in CFE XML doc
This PR aims to improve the way we create NomItem and DscItem lines on the XML files we send to DGI in order to generate the PDF report.

We have two fields to inform to DGI in an electronic invoice, NomItem which is mandatory an has a maximum of 80 characters to send, and DscItem that is optional an has a limit of 1000 chars. So we ensure NomItem is always sent, and the description will be filled like this:

First, if the product name is longer than 80 chars, DscItem field will contain the characters that didn't fit in the NomItem field (characters from the 80th)
Then it will contain the description of the product and the addenda content if it exists.
Plus, I refactor the method _l10n_uy_edi_cfe_D_global_discount so it is also aligned with what I mentioned before.

Another reason for this improvement is that, as the description of a product contains the product name at the beginning, we faced some issues when a user added a description on the fly in the line of the product, and then confirmed and printed the report. In this case, the name of the product was duplicated in the report. With this change, the description will only contain the the part of the product name that exceeds the 80th character.

closes odoo/enterprise#89440

X-original-commit: 67cd289eb3d0a0e388508de1b350b5ed5bbfe1a9
Signed-off-by: Claire Bretton (clbr) <clbr@odoo.com>