[IMP] knowledge: adapt comment plugin to toolbar changes
The following changes have been made in the toolbar in html_editor:
- the `namespace` parameter in the toolbar item definition has been
changed to `namespaces`, supporting multiple namespaces for a toolbar
group or item.
- the toolbar opens in a compact mode, with a small subset of buttons,
and can be expanded to display a larger set of buttons (namely those in
the "expanded" namespace).
- the `title` parameter has been replaced with `description`, which is
used to define the content of the tooltip displayed when hovering over
the item.

This commit adapts the Comments plugin to these changes.

task-4535289

closes odoo/enterprise#81075

Related: odoo/odoo#201037
Signed-off-by: David Monjoie (dmo) <dmo@odoo.com>