[FIX] pos_self_order_iot: updating screen orientation bug
We made invisible the `screen_orientation` field if `can_be_kiosk` was `False` on the IoT Box, but
we still updated the screen orientation if `screen_orientation` was set: which was always the
case as it has a default value.
We now update only if the IoT Box `can_be_kiosk`.

closes odoo/enterprise#81080

Task: 4606855
X-original-commit: 26b5cdf0e7b3fec2e8f38ab29c9e9edf1bb1e515
Signed-off-by: Yaroslav Soroko (yaso) <yaso@odoo.com>
Signed-off-by: Louis Travaux (lotr) <lotr@odoo.com>