[IMP] industry_fsm_sale: change label for task_id field in SO
Before this commit, the task_id is confusing for the user who would like
to create a custom filter to see the SO linked to a specific project, he
does not know if he can use that field or `task_ids` field. Actually,
`task_id` field is just used when `Extra Quotations` feature is enabled
and the user creates a quotation inside a task form view, which means
the field will just contain the task in which the quotation has been
created and that's it.

This commit changes the label of `task_id` to say the field will contain
the **source task**.

task-4623653

closes odoo/enterprise#81062

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>